More css moved to external file
